Move header files to header dir
