+++ components/common/header/authNav.js
@@ -1,7 +1,7 @@
 import { useState } from 'react'
 import { useRouter } from 'next/router'
 
-import { logout } from '../../client'
+import { logout } from '../../../client'
 
 import {
   useTheme,
