+++ components/common/header/fullLogo.js
