+++ components/common/header/logoLetters.js
